Refactor report structure and move components to ptrs directory
- Moved FinalReview component and its loader to ptrs directory.
- Deleted original FinalReview component and updated imports accordingly.
- Moved SectionForm component to ptrs directory and updated imports.
- Deleted original SectionForm component.
- Moved SectionReview component to ptrs directory and updated imports.
- Deleted original SectionReview component.
- Moved prepareReport function to ptrs directory and updated imports.
- Updated Dashboard component to import prepareReport from new location.
- Adjusted imports in AppRouter to reflect new file structure.
